# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/class-sections": {
      "get": {
        "operationId": "ClassSectionsController_findAll",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text search over course code and course name.",
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "academicYear",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 2000,
              "maximum": 2100,
              "type": "number"
            }
          },
          {
            "name": "term",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 3,
              "type": "number"
            }
          },
          {
            "name": "mine",
            "required": false,
            "in": "query",
            "description": "`?mine=true` keeps only the sections the caller owns.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ClassSectionDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid Core Hub session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The role or ownership does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Staff: list and search class sections",
        "tags": [
          "class-sections"
        ]
      },
      "post": {
        "operationId": "ClassSectionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassSectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ClassSectionDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid Core Hub session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The role or ownership does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Staff: create a class section - the caller becomes its owner",
        "tags": [
          "class-sections"
        ]
      }
    },
    "/api/v1/class-sections/{id}": {
      "get": {
        "operationId": "ClassSectionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ClassSectionDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid Core Hub session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The role or ownership does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Staff: one class section",
        "tags": [
          "class-sections"
        ]
      },
      "patch": {
        "operationId": "ClassSectionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassSectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ClassSectionDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid Core Hub session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The role or ownership does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Staff: change the name, location or thresholds of a section",
        "tags": [
          "class-sections"
        ]
      },
      "delete": {
        "operationId": "ClassSectionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedClassSectionDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid Core Hub session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The role or ownership does not allow this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Staff: delete a section that has no attendance history",
        "tags": [
          "class-sections"
        ]
      }
    },
...
  },
  "info": {
    "title": "Attendance Checker API",
    "description": "csmju-attendance-checker - CSMJU2030 subsystem. Every /api/v1 route needs the Core Hub access token: the HttpOnly session cookie set by /auth/callback, or a Bearer header.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "csmju_attendance_checker_access_token"
      }
    },
    "schemas": {
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "VALIDATION_ERROR",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "CONFLICT",
          "TOO_MANY_REQUESTS",
          "INTERNAL_ERROR",
          "SERVICE_UNAVAILABLE"
        ],
        "description": "One of the codes in standards/contracts/error-codes.json."
      },
      "ApiErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "description": "One of the codes in standards/contracts/error-codes.json.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "details": {
            "type": "object",
            "description": "Field errors for VALIDATION_ERROR, or extra facts such as `distanceMeters`."
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorBodyDto"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
      "ClassSectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "courseCode": {
            "type": "string"
          },
          "courseName": {
            "type": "string"
          },
          "sectionCode": {
            "type": "string"
          },
          "academicYear": {
            "type": "number",
            "description": "Gregorian year; the UI shows the Buddhist era."
          },
          "term": {
            "type": "number"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "radiusMeters": {
            "type": "number",
            "description": "Students must be within this many meters of the check-in point."
          },
          "ownerCoreUserId": {
            "type": "string",
            "description": "Core Hub user (`sub`) of the staff member who owns the section."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "courseCode",
          "courseName",
          "sectionCode",
          "academicYear",
          "term",
          "latitude",
          "longitude",
          "radiusMeters",
          "ownerCoreUserId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateClassSectionDto": {
        "type": "object",
        "properties": {
          "courseCode": {
            "type": "string",
            "description": "e.g. CS201",
            "pattern": "^[A-Z]{2,4}\\d{3,4}$"
          },
          "courseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "sectionCode": {
            "type": "string",
            "description": "e.g. 1 or 01",
            "pattern": "^\\d{1,3}$"
          },
          "academicYear": {
            "type": "number",
            "description": "Gregorian year, e.g. 2026 (the UI shows 2569).",
            "minimum": 2000,
            "maximum": 2100
          },
          "term": {
            "type": "number",
            "minimum": 1,
            "maximum": 3
          },
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "radiusMeters": {
            "type": "number",
            "minimum": 10,
            "maximum": 500
          }
        },
        "required": [
          "courseCode",
          "courseName",
          "sectionCode",
          "academicYear",
          "term",
          "latitude",
          "longitude"
        ]
      },
      "UpdateClassSectionDto": {
        "type": "object",
        "properties": {
          "courseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "radiusMeters": {
            "type": "number",
            "minimum": 10,
            "maximum": 500
          }
        }
      },
      "DeletedClassSectionDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "id"
        ]
      },
...
    }
  }
}